Add etcd-cfn-signal app. (#3)
* Add etcd-cfn-signal app.

Why do we need this change?
---

The old cfn-signal app did a good job of singalling to cloudformation
when the machine had finished starting up, so it could continue with a
rolling deploy.

However, it had one potential problem: it only waited for etcd to start
up, not for it to actually be ready.  This could lead to cloudformation
_thinking_ that the machines are all ready when all the new ones are
actually failing to join the existing cluster.  If this happened, it
would eventually replace all of our _working_ machines with broken ones.

Obviously this is not good.

What effects does this change have?
---

This adds a new app called `etcd-cfn-signal` that will specifically
check the health of the local etcd instance before signalling
cloudformation.

This should mean that any new instances don't get marked as good until
the point that they've actually joined the cluster.  In the event that a
machine fails to join the cluster the signal should never be sent, and
cloudformation will rollback to the old working config.  Even if that
rollback fails, we should at least not replace _all_ the machines that
have a working config.

Implementation Details
---

Decided to write this in rust because I wanted to get a bit of
expereience with it, and see how it compares to go.  Also, the
fact that it has a fairly modern type system felt like it would make
writing correct code a bit easier.  (No null pointer exceptions, yay!)

* PR fix

* More PR fixes

* More PR fixes