Add monitor-asg-lifecycle tool. (#1)
This adds a tool that will listen for SQS messages on the current
autoscale groups lifecycle hook queue.  When a message comes in to let
us know that a host has been terminate, it removes that host from the
etcd cluster.

This helps maintain quorum, as we don't consider dead hosts as part of
the cluster forever, which is the default etcd behaviour.